# docs/mobile-handoff/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MedicalPlab Product API",
    "description": "Fail-closed MedicalPlab API for pilot and production clients",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/adaptive/state": {
      "get": {
        "tags": [
          "Adaptive Learning Engine"
        ],
        "summary": "Get Learner State",
        "description": "Retrieve the unified learner state, topic mastery, weaknesses, and recommendations.\n\nEndpoint: GET /api/v1/adaptive/state\nPurpose: Retrieve unified learner state.\nInput: Header X-User-Id or Query learner_id.\nOutput: LearnerState.\nReason: Single unified source of truth on student educational progress.",
        "operationId": "get_learner_state_api_v1_adaptive_state_get",
        "parameters": [
          {
            "name": "learner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Learner Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-Learner-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Learner-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adaptive/recommendation": {
      "get": {
        "tags": [
          "Adaptive Learning Engine"
        ],
        "summary": "Get Recommendation",
        "description": "Retrieve the highest priority next learning action chosen by the Adaptive Decision Engine.\n\nEndpoint: GET /api/v1/adaptive/recommendation\nPurpose: Next best learning action.\nInput: Header X-User-Id or Query learner_id.\nOutput: AdaptiveRecommendation or null.\nReason: Guides student directly to their next intervention.",
        "operationId": "get_recommendation_api_v1_adaptive_recommendation_get",
        "parameters": [
          {
            "name": "learner_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Learner Id"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-Learner-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Learner-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AdaptiveRecommendation"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Recommendation Api V1 Adaptive Recommendation Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adaptive/event": {
      "post": {
        "tags": [
          "Adaptive Learning Engine"
        ],
        "summary": "Record Learning Event",
        "description": "Ingest a learning interaction event and return the updated learner state.\n\nEndpoint: POST /api/v1/adaptive/event\nPurpose: Ingest learning interaction event.\nInput: LearningEvent JSON body.\nOutput: Updated LearnerState.\nReason: Enables real-time mastery recalculation following practice or review.",
        "operationId": "record_learning_event_api_v1_adaptive_event_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-Learner-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Learner-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearningEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adaptive/remediate": {
      "post": {
        "tags": [
          "Adaptive Learning Engine"
        ],
        "summary": "Trigger Remediation",
        "description": "Trigger an evidence-grounded Socratic tutor intervention targeted at student weakness.\n\nEndpoint: POST /api/v1/adaptive/remediate\nPurpose: Trigger Grounded Tutor remediation for active weakness.\nInput: RemediationRequest JSON body.\nOutput: TutorChatResponse (verified by Phase 1 verifier).\nReason: Connects adaptive gap detection to verified Socratic tutor safely.",
        "operationId": "trigger_remediation_api_v1_adaptive_remediate_post",
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-Learner-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Learner-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemediationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/adaptive/loop-status": {
      "get": {
        "tags": [
          "Adaptive Learning Engine"
        ],
        "summary": "Get Loop Status",
        "description": "Check whether mastery has improved following an intervention.\n\nEndpoint: GET /api/v1/adaptive/loop-status\nPurpose: Evaluate learning loop closure.\nInput: Query params topic, baseline_mastery, Header X-User-Id.\nOutput: JSON summary of initial vs current mastery.\nReason: Closes feedback loop across diagnostic assessment and re-evaluation.",
        "operationId": "get_loop_status_api_v1_adaptive_loop_status_get",
        "parameters": [
          {
            "name": "topic",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Topic"
            }
          },
          {
            "name": "baseline_mastery",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Baseline Mastery"
            }
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "name": "X-Learner-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Learner-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Loop Status Api V1 Adaptive Loop Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdaptiveActionType": {
        "type": "string",
        "enum": [
          "SOLVE_TARGETED_QUESTION",
          "ASK_GROUNDED_TUTOR",
          "REVIEW_CONCEPT",
          "REPEAT_TOPIC"
        ],
        "title": "AdaptiveActionType"
      },
      "AdaptiveRecommendation": {
        "properties": {
          "recommendation_id": {
            "type": "string",
            "title": "Recommendation Id"
          },
          "action": {
            "$ref": "#/components/schemas/AdaptiveActionType"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "priority": {
            "$ref": "#/components/schemas/WeaknessPriority"
          },
          "target_question_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Question Id"
          },
          "tutor_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tutor Mode"
          },
          "tutor_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tutor Query"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "recommendation_id",
          "action",
          "subject",
          "topic",
          "priority",
          "reason",
          "explanation"
        ],
        "title": "AdaptiveRecommendation",
        "description": "Next best pedagogical action chosen by the decision engine."
      },
...
      "AttemptRecord": {
        "properties": {
          "attempt_key": {
            "type": "string",
            "title": "Attempt Key"
          },
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "selected": {
            "type": "string",
            "title": "Selected"
          },
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "attempt_key",
          "question_id",
          "subject",
          "topic",
          "selected",
          "correct"
        ],
        "title": "AttemptRecord",
        "description": "Normalized representation of a single learner attempt."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LearnerState": {
        "properties": {
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "total_attempts": {
            "type": "integer",
            "title": "Total Attempts",
            "default": 0
          },
          "correct_attempts": {
            "type": "integer",
            "title": "Correct Attempts",
            "default": 0
          },
          "overall_accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Accuracy"
          },
          "topic_mastery": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TopicMasteryRecord"
            },
            "type": "object",
            "title": "Topic Mastery"
          },
          "weak_topics": {
            "items": {
              "$ref": "#/components/schemas/WeakTopicRecord"
            },
            "type": "array",
            "title": "Weak Topics"
          },
          "recent_activity": {
            "items": {
              "$ref": "#/components/schemas/AttemptRecord"
            },
            "type": "array",
            "title": "Recent Activity"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/AdaptiveRecommendation"
            },
            "type": "array",
            "title": "Recommendations"
          }
        },
        "type": "object",
        "required": [
          "learner_id"
        ],
        "title": "LearnerState",
        "description": "Unified educational representation of the learner."
      },
      "LearningEvent": {
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "learner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learner Id"
          },
          "event_type": {
            "type": "string",
            "enum": [
              "QUESTION_ATTEMPT",
              "TUTOR_INTERVENTION",
              "TOPIC_REVIEW"
            ],
            "title": "Event Type",
            "default": "QUESTION_ATTEMPT"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "question_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "selected_option": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Option"
          },
          "is_correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Correct"
          },
          "attempt_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempt Key"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "topic"
        ],
        "title": "LearningEvent",
        "description": "Incoming learning telemetry event."
      },
...
      "RemediationRequest": {
        "properties": {
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "question_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question Id"
          },
          "preferred_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Mode"
          },
          "custom_query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Query"
          }
        },
        "type": "object",
        "title": "RemediationRequest",
        "description": "Request to generate an evidence-grounded tutor intervention for a weakness."
      },
...
      "TopicMasteryRecord": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "total_attempts": {
            "type": "integer",
            "title": "Total Attempts",
            "default": 0
          },
          "correct_attempts": {
            "type": "integer",
            "title": "Correct Attempts",
            "default": 0
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "mastery_score": {
            "type": "number",
            "title": "Mastery Score",
            "default": 0.0
          },
          "mastery_level": {
            "$ref": "#/components/schemas/MasteryLevel",
            "default": "beginner"
          },
          "mastery_status": {
            "$ref": "#/components/schemas/MasteryStatus",
            "default": "not_started"
          },
          "confidence": {
            "$ref": "#/components/schemas/ConfidenceLevel",
            "default": "low"
          },
          "consecutive_errors": {
            "type": "integer",
            "title": "Consecutive Errors",
            "default": 0
          },
          "last_attempt_correct": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt Correct"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "topic"
        ],
        "title": "TopicMasteryRecord",
        "description": "Deterministic mastery status for an individual subject/topic."
      },
...
      "TutorChatResponse": {
        "properties": {
          "response_id": {
            "type": "string",
            "title": "Response Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "socratic_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Socratic Question"
          },
          "hints": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hints"
          },
          "misconception": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Misconception"
          },
          "mechanistic_explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mechanistic Explanation"
          },
          "distractor_analysis": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DistractorItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distractor Analysis"
          },
          "revision_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision Summary"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/TutorCitationDTO"
            },
            "type": "array",
            "title": "Citations"
          },
          "evidence_packet_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Packet Id"
          },
          "support_status": {
            "type": "string",
            "enum": [
              "SUPPORTED",
              "SAFE_FALLBACK",
              "PARTIALLY_SUPPORTED",
              "UNSUPPORTED",
              "ABSTAIN"
            ],
            "title": "Support Status",
            "default": "SUPPORTED"
          },
          "abstain": {
            "type": "boolean",
            "title": "Abstain",
            "default": false
          },
          "abstain_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstain Reason"
          },
          "fallback_applied": {
            "type": "boolean",
            "title": "Fallback Applied",
            "default": false
          },
          "pedagogical_state": {
            "type": "string",
            "enum": [
              "PRE_SUBMISSION",
              "POST_SUBMISSION",
              "GENERAL_STUDY"
            ],
            "title": "Pedagogical State",
            "default": "PRE_SUBMISSION"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "stub"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": "stub"
          },
          "requested_provider": {
            "type": "string",
            "title": "Requested Provider",
            "default": "stub"
          },
          "requested_model": {
            "type": "string",
            "title": "Requested Model",
            "default": "stub"
          },
          "actual_generation_provider": {
            "type": "string",
            "title": "Actual Generation Provider",
            "default": "stub"
          },
          "actual_generation_model": {
            "type": "string",
            "title": "Actual Generation Model",
            "default": "stub"
          },
          "provider_failover_applied": {
            "type": "boolean",
            "title": "Provider Failover Applied",
            "default": false
          },
          "primary_provider_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Provider Error"
          },
          "latency_breakdown": {
            "$ref": "#/components/schemas/LatencyBreakdownDTO"
          },
          "verification": {
            "$ref": "#/components/schemas/TutorVerificationSummaryDTO"
          },
          "draft_verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TutorVerificationSummaryDTO"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "response_id",
          "session_id",
          "mode",
          "message"
        ],
        "title": "TutorChatResponse"
      },
...
      "WeakTopicRecord": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "priority": {
            "$ref": "#/components/schemas/WeaknessPriority"
          },
          "failure_rate": {
            "type": "number",
            "title": "Failure Rate"
          },
          "missed_count": {
            "type": "integer",
            "title": "Missed Count"
          },
          "total_attempts": {
            "type": "integer",
            "title": "Total Attempts"
          },
          "consecutive_errors": {
            "type": "integer",
            "title": "Consecutive Errors",
            "default": 0
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "topic",
          "priority",
          "failure_rate",
          "missed_count",
          "total_attempts",
          "reason"
        ],
        "title": "WeakTopicRecord",
        "description": "Identified knowledge deficit in a specific medical topic."
      },
...
    }
  }
}